% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{arrange_by}
\alias{arrange_by}
\title{Organise literature data by selected column}
\usage{
arrange_by(df, selection, sep = ";")
}
\arguments{
\item{df}{A data frame containing the literature data}

\item{selection}{A string with a column name}

\item{split}{A character used to split selection}
}
\value{
A data frame with literature organised by selected column
}
\description{
Organise literature data by selected column
}
