% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{build_stm_corpus}
\alias{build_stm_corpus}
\title{Create stm library corpus}
\usage{
build_stm_corpus(textlist)
}
\arguments{
\item{textlist}{Vector of processed words}
}
\value{
Corpus for stm
}
\description{
Create stm library corpus
}
