% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{build_topicmodel_from_literature}
\alias{build_topicmodel_from_literature}
\title{Create a topicmodel from WoS library data frame}
\usage{
build_topicmodel_from_literature(literature, K)
}
\arguments{
\item{literature}{Cleaned and processed literature dataframe}

\item{K}{Predefined K. If null, auto select K with stm library.}
}
\value{
List with topicmodels fit (fit), thetaDF distributions (thetadf) and subset of literature columns with topicmodel columns added (doctablewt)
}
\description{
Create a topicmodel from WoS library data frame
}
