% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{build_topicmodels_corpus}
\alias{build_topicmodels_corpus}
\title{Create topicmodels library corpus}
\usage{
build_topicmodels_corpus(textlist)
}
\arguments{
\item{textlist}{Vector of texts}
}
\value{
Corpus for topicmodels
}
\description{
Create topicmodels library corpus
}
