% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{change_uppercase}
\alias{change_uppercase}
\title{Helper function to change strings to uppercase in
selected data frame columns.}
\usage{
change_uppercase(df, columns)
}
\arguments{
\item{df}{A data frame}

\item{columns}{A vector of column names to be processes}
}
\value{
A data frame
}
\description{
Helper function to change strings to uppercase in
selected data frame columns.
}
