% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{check_data}
\alias{check_data}
\title{Check the data frame columns.}
\usage{
check_data(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
An integer from 1 to 3
}
\description{
Check the data frame columns.
}
