% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_author_network.R
\name{collapse_edge}
\alias{collapse_edge}
\title{Helper function for pasting two nodes together}
\usage{
collapse_edge(edge)
}
\arguments{
\item{node}{A list containing two author network nodes}
}
\value{
A string containing two nodes separated by ";"
}
\description{
Helper function for pasting two nodes together
}
