% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_author_network.R
\name{create_edges}
\alias{create_edges}
\title{Helper function for creating a node from author names}
\usage{
create_edges(x)
}
\arguments{
\item{x}{A list of nodes}
}
\value{
A list of edges
}
\description{
Helper function for creating a node from author names
}
