% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{fix_column_names}
\alias{fix_column_names}
\title{Fix column names of dataset from Web of Science}
\usage{
fix_column_names(column_names)
}
\arguments{
\item{column_names}{Column names}
}
\value{
Fixed column names
}
\description{
Fix column names of dataset from Web of Science
}
