% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_author_network.R
\name{get_address}
\alias{get_address}
\title{Helper function for extracting author address}
\usage{
get_address(df)
}
\arguments{
\item{df}{A row from a data frame}
}
\value{
A string containing author's address, if one exists
}
\description{
Helper function for extracting author address
}
