% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_author_network.R
\name{get_author_edges}
\alias{get_author_edges}
\title{Extract edges for an author network from literature data}
\usage{
get_author_edges(df)
}
\arguments{
\item{df}{a data frame containing literature data}
}
\value{
A data frame containing nodes for an author network
}
\description{
Extract edges for an author network from literature data
}
