% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_author_network.R
\name{get_author_network}
\alias{get_author_network}
\title{Extract author network nodes and edges from literature data}
\usage{
get_author_network(df, as_igraph = FALSE)
}
\arguments{
\item{df}{A data frame containing literature data}

\item{as_igraph}{Logical: should results be returned as an igraph?}
}
\value{
A list containing nodes and edges in data frames, or an igraph.
}
\description{
Extract author network nodes and edges from literature data
}
