% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{get_citation_edges}
\alias{get_citation_edges}
\title{Extract edges for a citation network from reference list}
\usage{
get_citation_edges(reference_list)
}
\arguments{
\item{reference_list}{A data frame containing reference list}
}
\value{
A data frame containing edges for a citation network
}
\description{
Extract edges for a citation network from reference list
}
