% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_citation_network.R
\name{get_citation_nodes}
\alias{get_citation_nodes}
\title{Extract nodes for a citation network from literature data}
\usage{
get_citation_nodes(df, reference_list)
}
\arguments{
\item{df}{A data frame containing literature data}

\item{reference_list}{A data frame containing reference list}
}
\value{
A data frame containing nodes for a citation network
}
\description{
Extract nodes for a citation network from literature data
}
