% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{get_DOI}
\alias{get_DOI}
\title{Helper function to extract DOIs}
\usage{
get_DOI(x)
}
\arguments{
\item{x}{A list of strings containing DOI in second element}
}
\value{
A string cointaining DOI
}
\description{
Helper function to extract DOIs
}
