% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_citation_network.R
\name{get_reference_list}
\alias{get_reference_list}
\title{Helper function for extracting reference lists}
\usage{
get_reference_list(df)
}
\arguments{
\item{df}{A data frame containing literature data}
}
\value{
A data frame containing each reference in separate row
}
\description{
Helper function for extracting reference lists
}
