% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{get_year}
\alias{get_year}
\title{Helper function to extract years}
\usage{
get_year(x)
}
\arguments{
\item{x}{A list of strings containing year in the second element}
}
\value{
A year number
}
\description{
Helper function to extract years
}
