% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{getName}
\alias{getName}
\title{Helper function to extract the name of first author}
\usage{
getName(x)
}
\arguments{
\item{x}{A string containing full author names}
}
\value{
A string
}
\description{
Helper function to extract the name of first author
}
