% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{make_reference}
\alias{make_reference}
\title{Helper function to construct reference strings}
\usage{
make_reference(x)
}
\arguments{
\item{x}{A row from a data frame containing cleaned data from WOS}
}
\value{
A string in same format as references in CitedReferences
}
\description{
Helper function to construct reference strings
}
