% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_wos_data}
\alias{read_wos_data}
\title{Read file or folder of files downloaded from Web of Science}
\usage{
read_wos_data(filepath)
}
\arguments{
\item{filepath}{Path to a file or folder containting the data}
}
\value{
A data frame
}
\description{
Read file or folder of files downloaded from Web of Science
}
