% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{remove_quotes}
\alias{remove_quotes}
\title{Helper function to remove quotation marks from strings.}
\usage{
remove_quotes(df, columns)
}
\arguments{
\item{df}{A data frame}

\item{columns}{A vector of column names to be processed}
}
\value{
A data frame
}
\description{
Helper function to remove quotation marks from strings.
}
