% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{select_optimal_k}
\alias{select_optimal_k}
\title{Estimates number of topics (K) using the stm library}
\usage{
select_optimal_k(data)
}
\arguments{
\item{data}{Data from preprocess_literature_for_topicmodeling()}
}
\value{
A list with topic estimation from stm lib (topicest), semantic coherence values (semcohvalues) and estimate of best K of topics (bestk)
}
\description{
Estimates number of topics (K) using the stm library
}
