% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{topicmodels_json_ldavis}
\alias{topicmodels_json_ldavis}
\title{Convert the output of a topicmodels Latent Dirichlet Allocation to JSON
for use with LDAvis}
\usage{
topicmodels_json_ldavis(fitted, corpus, doc_term)
}
\arguments{
\item{fitted}{Output from a topicmodels \code{LDA} model.}

\item{corpus}{Corpus object used to create the document term
matrix for the \code{LDA} model. This should have been create with
the tm package's \code{Corpus} function.}

\item{doc_term}{The document term matrix used in the \code{LDA}
model. This should have been created with the tm package's
\code{DocumentTermMatrix} function.}
}
\description{
Convert the output of a topicmodels Latent Dirichlet Allocation to JSON
for use with LDAvis
}
\seealso{
\link{LDAvis}.
}
