% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fieldtags.R
\name{update_fieldtags}
\alias{update_fieldtags}
\title{Update internally saved field tags}
\usage{
update_fieldtags()
}
\value{
Announcement that fieldtags have been updated.
}
\description{
Update internally saved field tags
}
